Update package versions and refactor audit logic
Updated "Remora.Discord" package version to 2023.3.0 to acquire new features and fixes. Replaced "ClientStatus" to "UserStatus" enumeration to align with revised library definitions in ReadyResponder.cs. Refactored getting audit logs in MemberUpdateResponder.cs to use the updated method call and simplified the condition to get the most recent timeout event. These updates improve compatibility with the newer version of the library and streamline the audit log process for better performance.