Refactor condition checks and remove deconstruct handle
Simplified and consolidated the conditional check in the MemberUpdateResponder.cs for better readability, and improved efficiency. The redundancies in DiscordHandleHelper.cs were eliminated by removing the deconstruct handle function as it was unused and not necessary. A condition check was added for when the discriminator is zero to just return the username.